refactor(hub): address review — schemas, lifecycle, one error type
- Schemas move out of the routers into `_router/_schema/`, matching
  every other router in the tree: `_hub.py`, `_hub_mcp.py`,
  `_hub_skill.py`, `_workspace.py`.

- Hubs become async context managers, entered by the app lifespan. Each
  now holds one HTTP client for the life of the process rather than
  handshaking afresh on every catalog page. A download's client
  outlives the response it streams, so only the response is closed when
  the body is drained.

- `ClawHubError` and `GitHubMCPError` collapse into one `HubError`
  carrying the `hub_id`. They only ever differed by name, and nothing
  caught either — which is also why a hub's failure used to surface as
  a 500. An exception handler now maps it to 503 for a rate limit and
  502 otherwise: a third party's failure is not this service's.

- `render_mcp` moves to the service layer as `_service/_mcp_render.py`.

- `resolve_workspace` goes back to being private in the workspace
  router. It was lifted into `deps.py` for a hub-router caller that no
  longer exists, and it never was a dependency — it takes five
  arguments and is called by hand.

- Storage settles on one naming scheme: `list_skills` / `get_skill` /
  `get_skill_by_name` / `delete_skill`, matching the MCP side.

- `_skill_upload`'s module-level functions take the `_` prefix the
  repo uses for package-private helpers.

- The GitHub hub inlines two single-use helpers; the example MCP hub
  (`StaticMCPHub`) is dropped along with a stray blank line in the web
  client.
