+++ src/agentscope/app/_router/_schema/__init__.py
@@ -2,6 +2,17 @@
 """Schema models for the agent service."""
 
 from ._chat import ChatRequest, ChatTriggerResponse
+from ._hub import HubInfo
+from ._hub_mcp import InstallMCPRequest, MCPView, UpdateMCPRequest
+from ._hub_skill import SkillView
+from ._workspace import (
+    AddFromLibraryRequest,
+    AddFromLibraryResponse,
+    AddSkillRequest,
+    AddSkillsFromLibraryRequest,
+    MCPClientStatus,
+    ToolInfo,
+)
 from ._model import ListModelsResponse, ListModelsRequest
 from ._tts_model import ListTTSModelsResponse, ListTTSModelsRequest
 from ._schedule import (
@@ -57,6 +68,19 @@ from ._session import (
 )
 
 __all__ = [
+    # Hub
+    "HubInfo",
+    "InstallMCPRequest",
+    "MCPView",
+    "UpdateMCPRequest",
+    "SkillView",
+    # Workspace
+    "AddFromLibraryRequest",
+    "AddFromLibraryResponse",
+    "AddSkillRequest",
+    "AddSkillsFromLibraryRequest",
+    "MCPClientStatus",
+    "ToolInfo",
     # Agent
     "AgentSchemaResponse",
     "AgentSchemaV2Response",
