+++ src/agentscope/app/_service/__init__.py
@@ -12,6 +12,7 @@ from ._index_sweeper import IndexSweeper
 from ._index_task_consumer import IndexTaskConsumer
 from ._index_worker import IndexWorker
 from ._knowledge_base import KnowledgeBaseService
+from ._mcp_render import MCPRenderError, render_mcp
 from ._model import get_model
 from ._tts_model import get_tts_model
 from ._session import SessionService, SessionStatus
@@ -28,6 +29,7 @@ __all__ = [
     "IndexWorker",
     "KnowledgeBaseService",
     "KnowledgeBaseView",
+    "MCPRenderError",
     "ResourceAccessService",
     "SessionService",
     "SessionStatus",
@@ -37,4 +39,5 @@ __all__ = [
     "get_model",
     "get_tts_model",
     "get_toolkit",
+    "render_mcp",
 ]
