+++ src/agentscope/app/hub/_mcp/__init__.py
@@ -3,15 +3,11 @@
 
 from ._base import MCPHubBase
 from ._card import MCPCard, MCPHubPage
-from ._github_hub import GitHubMCPError, GitHubMCPHub
-from ._render import MCPRenderError, render_mcp
+from ._github_hub import GitHubMCPHub
 
 __all__ = [
-    "GitHubMCPError",
     "GitHubMCPHub",
     "MCPHubBase",
     "MCPCard",
     "MCPHubPage",
-    "MCPRenderError",
-    "render_mcp",
 ]
