+++ tests/workspace_skill_archive_test.py
@@ -13,8 +13,8 @@ from fastapi import UploadFile
 from agentscope.app._service._skill_upload import (
     SkillUploadError,
     UploadManifest,
-    tar_stream,
-    validate_manifest,
+    _tar_stream,
+    _validate_manifest,
 )
 from agentscope.workspace import LocalWorkspace
 from agentscope.workspace._base import WorkspaceBase
@@ -222,25 +222,25 @@ class UploadManifestTest(IsolatedAsyncioTestCase):
         manifest = self._manifest(
             {"pack/SKILL.md": b"x", "pack/lib/run.py": b"y"},
         )
-        self.assertEqual(validate_manifest(manifest), "pack")
+        self.assertEqual(_validate_manifest(manifest), "pack")
 
     def test_rejects_multiple_roots(self) -> None:
         """Files from two folders are not one skill."""
         manifest = self._manifest({"a/SKILL.md": b"x", "b/run.py": b"y"})
         with self.assertRaises(SkillUploadError):
-            validate_manifest(manifest)
+            _validate_manifest(manifest)
 
     def test_rejects_traversal(self) -> None:
         """A ``..`` segment is refused before anything is read."""
         manifest = self._manifest({"pack/../../SKILL.md": b"x"})
         with self.assertRaises(SkillUploadError):
-            validate_manifest(manifest)
+            _validate_manifest(manifest)
 
     def test_rejects_missing_skill_md(self) -> None:
         """The root must hold a SKILL.md."""
         manifest = self._manifest({"pack/readme.md": b"x"})
         with self.assertRaises(SkillUploadError):
-            validate_manifest(manifest)
+            _validate_manifest(manifest)
 
     def test_rejects_oversized_file(self) -> None:
         """A single part over the per-file limit is refused."""
@@ -253,7 +253,7 @@ class UploadManifestTest(IsolatedAsyncioTestCase):
             },
         )
         with self.assertRaises(SkillUploadError):
-            validate_manifest(manifest)
+            _validate_manifest(manifest)
 
     async def test_stream_round_trips(self) -> None:
         """The emitted tar reads back with the same members."""
@@ -263,7 +263,8 @@ class UploadManifestTest(IsolatedAsyncioTestCase):
         }
         manifest = self._manifest(files)
         chunks = [
-            chunk async for chunk in tar_stream(manifest, self._uploads(files))
+            chunk
+            async for chunk in _tar_stream(manifest, self._uploads(files))
         ]
 
         archive = tarfile.open(fileobj=io.BytesIO(b"".join(chunks)))
@@ -282,5 +283,5 @@ class UploadManifestTest(IsolatedAsyncioTestCase):
         )
         uploads = self._uploads({"pack/SKILL.md": b"much longer"})
         with self.assertRaises(SkillUploadError):
-            async for _ in tar_stream(manifest, uploads):
+            async for _ in _tar_stream(manifest, uploads):
                 pass
